+++ packages/core/vitest/matterbridgeEndpoint-default.test.ts
@@ -2507,13 +2507,14 @@ describe('Matterbridge ' + NAME, () => {
     const device = new MatterbridgeEndpoint(temperatureSensor, { id: 'TemperatureSensor' });
     expect(device).toBeDefined();
     device.createDefaultIdentifyClusterServer();
-    device.createDefaultTemperatureMeasurementClusterServer(21 * 100);
+    device.createDefaultTemperatureMeasurementClusterServer(21 * 100, null, null, 10);
     expect(device.hasClusterServer(TemperatureMeasurement.id)).toBe(true);
     expect(device.hasAttributeServer(TemperatureMeasurement.id, 'measuredValue')).toBe(true);
 
     await add(device);
 
     expect(device.getAttribute(TemperatureMeasurement.id, 'measuredValue')).toBe(2100);
+    expect(device.getAttribute(TemperatureMeasurement.id, 'tolerance')).toBe(10);
     // (matterbridge.frontend as any).getClusterTextFromDevice(device);
   });
 
@@ -2521,13 +2522,14 @@ describe('Matterbridge ' + NAME, () => {
     const device = new MatterbridgeEndpoint(humiditySensor, { id: 'HumiditySensor' });
     expect(device).toBeDefined();
     device.createDefaultIdentifyClusterServer();
-    device.createDefaultRelativeHumidityMeasurementClusterServer(50 * 100);
+    device.createDefaultRelativeHumidityMeasurementClusterServer(50 * 100, null, null, 20);
     expect(device.hasClusterServer(RelativeHumidityMeasurement.id)).toBe(true);
     expect(device.hasAttributeServer(RelativeHumidityMeasurement.id, 'measuredValue')).toBe(true);
 
     await add(device);
 
     expect(device.getAttribute(RelativeHumidityMeasurement.id, 'measuredValue')).toBe(5000);
+    expect(device.getAttribute(RelativeHumidityMeasurement.id, 'tolerance')).toBe(20);
     // (matterbridge.frontend as any).getClusterTextFromDevice(device);
   });
 
@@ -2535,13 +2537,14 @@ describe('Matterbridge ' + NAME, () => {
     const device = new MatterbridgeEndpoint(pressureSensor, { id: 'PressureSensor' });
     expect(device).toBeDefined();
     device.createDefaultIdentifyClusterServer();
-    device.createDefaultPressureMeasurementClusterServer(980);
+    device.createDefaultPressureMeasurementClusterServer(980, null, null, 30);
     expect(device.hasClusterServer(PressureMeasurement.id)).toBe(true);
     expect(device.hasAttributeServer(PressureMeasurement.id, 'measuredValue')).toBe(true);
 
     await add(device);
 
     expect(device.getAttribute(PressureMeasurement.id, 'measuredValue')).toBe(980);
+    expect(device.getAttribute(PressureMeasurement.id, 'tolerance')).toBe(30);
     // (matterbridge.frontend as any).getClusterTextFromDevice(device);
   });
 
@@ -2563,13 +2566,14 @@ describe('Matterbridge ' + NAME, () => {
     const device = new MatterbridgeEndpoint(flowSensor, { id: 'FlowSensor' });
     expect(device).toBeDefined();
     device.createDefaultIdentifyClusterServer();
-    device.createDefaultFlowMeasurementClusterServer(20 * 10);
+    device.createDefaultFlowMeasurementClusterServer(20 * 10, null, null, 40);
     expect(device.hasClusterServer(FlowMeasurement.id)).toBe(true);
     expect(device.hasAttributeServer(FlowMeasurement.id, 'measuredValue')).toBe(true);
 
     await add(device);
 
     expect(device.getAttribute(FlowMeasurement.id, 'measuredValue')).toBe(200);
+    expect(device.getAttribute(FlowMeasurement.id, 'tolerance')).toBe(40);
     // (matterbridge.frontend as any).getClusterTextFromDevice(device);
   });